Refactored DisplayListing
