+++ app/src/main/java/com/example/divvy/models/ViewHolderFactory.java
@@ -1,5 +1,6 @@
 package com.example.divvy.models;
 
+import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -7,11 +8,9 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.example.divvy.Controllers.ImageSelector;
 import com.example.divvy.R;
 
-import org.w3c.dom.Text;
-
-import java.util.List;
 
 public class ViewHolderFactory {
 
@@ -32,9 +31,15 @@ public class ViewHolderFactory {
 
         public ImageMessageViewHolder(View view){
             super(view);
+            message = view.findViewById(R.id.message);
+            username = view.findViewById(R.id.username);
+            image = view.findViewById(R.id.imageView);
         }
         @Override public void setUpUi(Object imageMessage){
-
+            ImageMessage message = (ImageMessage) imageMessage;
+            this.username.setText(message.getSender());
+            this.message.setText(message.getMessage());
+            this.image.setImageDrawable(ImageSelector.LoadImageFromWebOperations(""));
         }
     }
 
@@ -48,9 +53,26 @@ public class ViewHolderFactory {
         }
 
         @Override public void setUpUi(Object o){
-           Message message = (Message) o;
+            Message message = (Message) o;
             this.username.setText(message.getSender());
             this.message.setText(message.getMessage());
+        }
+    }
+
+    public static class ListingListViewHolder extends MyViewHolder {
+        TextView title;
+        TextView owner;
+
+        public ListingListViewHolder(@NonNull View listing) {
+            super(listing);
+            title = listing.findViewById(R.id.listing_frag_title);
+            owner = listing.findViewById(R.id.listing_frag_owner);
+        }
+
+        @Override public void setUpUi(Object o){
+            Listing listing = (Listing) o;
+            this.title.setText(listing.title);
+            this.owner.setText(listing.owner);
         }
     }
 
@@ -60,8 +82,11 @@ public class ViewHolderFactory {
             return new MessageViewHolder(LayoutInflater.from(parent.getContext())
                     .inflate(R.layout.message, parent, false));
         }else if(object.getClass().equals(ImageMessage.class)){
-            return new MessageViewHolder(LayoutInflater.from(parent.getContext())
+            return new ImageMessageViewHolder(LayoutInflater.from(parent.getContext())
                     .inflate(R.layout.message, parent, false));
+        }else if(object.getClass().equals(Listing.class)){
+            return new ListingListViewHolder(LayoutInflater.from(parent.getContext())
+                    .inflate(R.layout.fragment_listing, parent, false));
         }
         return null;
     }
